Update README.md
Discreption: Maching Memory is a game in which the player turns over face down cards and tries to find matching pairs. The type of cards in the game are card game cards and are distributed randomly These are the cards located on the cells of the game board. The game allows the player to set up a game timeout. The game displays a countdown timer on the screen/display while the game is running The game displays a game over message if the countdown timer reaches zero and the cards are not matched correctly. The player is able to select two cells, trying to find a matching pair. The game must face up to the face-up cards that the player hosts within the designated cells. The game must keep matching cards face up until the entire deck of cards is completed They match each other. Issues: A timer appears in reverse, i.e. 10, 9, 8...0, in a separate form from the game, and the warning message that the game has ended also appears in a separate form. Notes: When downloading the game code to the device, you must take into account the location of its storage and change it in the part related to images in the game code Team: Fayrouz Taha Samah Ali
it's a one version that has two frames a frame for the game and a frame for the countdown timer and wrong message.